Run full tree first to possibly short-circut exploration
When looking at at a benchmark where inlining is siginficant, we
grow each method from a noinline to a full tree, measuring each
step along the way. This is wasteful if that method itself is either
not called, not important for perf, or does not have inlines that
have perf impacts.

So instead, measure the full tree first, and compare the overall
perf versus noinline, and then only explore the subtree if the full
tree has significant perf impact over noinline.